util: improve stack message in the panic log
