+++ testkit/asynctestkit.go
@@ -19,13 +19,13 @@ package testkit
 import (
 	"context"
 	"fmt"
+	"runtime"
 	"testing"
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/tidb/kv"
 	"github.com/pingcap/tidb/session"
 	"github.com/pingcap/tidb/types"
-	"github.com/pingcap/tidb/util"
 	"github.com/pingcap/tidb/util/sqlexec"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -69,6 +69,15 @@ func (tk *AsyncTestKit) CloseSession(ctx context.Context) {
 	se.Close()
 }
 
+// GetStack gets the stacktrace.
+func GetStack() []byte {
+	const size = 4096
+	buf := make([]byte, size)
+	stackSize := runtime.Stack(buf, false)
+	buf = buf[:stackSize]
+	return buf
+}
+
 // ConcurrentRun run test in current.
 // - concurrent: controls the concurrent worker count.
 // - loops: controls run test how much times.
@@ -95,7 +104,7 @@ func (tk *AsyncTestKit) ConcurrentRun(
 		go func() {
 			defer func() {
 				r := recover()
-				tk.require.Nil(r, string(util.GetStack()))
+				tk.require.Nil(r, string(GetStack()))
 				doneList[w]()
 			}()
